Refactor: change module directory
