--- src/main/java/code/rice/bowl/spaghetti/config/JsonNodeConverter.java
@@ -1,31 +1,0 @@
-package code.rice.bowl.spaghetti.config;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
-
-@Converter(autoApply = false)
-public class JsonNodeConverter implements AttributeConverter<JsonNode, String> {
-
-    private static final ObjectMapper objectMapper = new ObjectMapper();
-
-    @Override
-    public String convertToDatabaseColumn(JsonNode attribute) {
-        try {
-            return objectMapper.writeValueAsString(attribute);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("JSON converting failed", e);
-        }
-    }
-
-    @Override
-    public JsonNode convertToEntityAttribute(String dbData) {
-        try {
-            return objectMapper.readTree(dbData);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("JSON reverse converting failed", e);
-        }
-    }
-}
